Add csv_file module
